+++ java/flight/flight-jdbc-driver/src/test/java/org/apache/arrow/driver/jdbc/utils/FlightStreamQueueTest.java
@@ -24,6 +24,7 @@ import static org.mockito.Mockito.mock;
 import java.util.concurrent.CompletionService;
 
 import org.apache.arrow.flight.FlightStream;
+import org.apache.arrow.flight.FlightStreamImpl;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -64,7 +65,7 @@ public class FlightStreamQueueTest {
   public void testEnqueueShouldThrowExceptionUponClose() throws Exception {
     queue.close();
     ThrowableAssertionUtils.simpleAssertThrowableClass(IllegalStateException.class,
-        () -> queue.enqueue(mock(FlightStream.class)));
+        () -> queue.enqueue(mock(FlightStreamImpl.class)));
   }
 
   @Test
